add folders, refactor the code
