+++ src/components/ui/MobileBtn/MobileBtn.tsx
@@ -1,14 +1,13 @@
 "use client";
-import React from "react";
+
+import { FC } from "react";
+
+import { MobileBtnType } from "@/components/ui/MobileBtn/types";
+
 import CloseIcon from "@/public/assets/icons/icon-close.svg";
 import MenuIcon from "@/public/assets/icons/icon-menu.svg";
 
-type MobileBtnProps = {
-  isMobileMenuShown: boolean;
-  handleClick: () => void;
-};
-
-const MobileBtn = ({ isMobileMenuShown, handleClick }: MobileBtnProps) => {
+const MobileBtn: FC<MobileBtnType> = ({ isMobileMenuShown, handleClick }) => {
   return (
     <button className="w-[24px] h-[24px] block lg:hidden" onClick={handleClick}>
       {!isMobileMenuShown ? <MenuIcon /> : <CloseIcon />}
