+++ src/components/ui/MobileBtn/types.ts
@@ -1,0 +1,4 @@
+export type MobileBtnType = {
+  isMobileMenuShown: boolean;
+  handleClick: () => void;
+};
