+++ src/components/ui/NavBar/NavBar.tsx
@@ -1,23 +1,17 @@
+import { FC } from "react";
 import { Link } from "react-scroll";
 
-import React from "react";
-import data from "@/data/navbar.json";
-type navData = {
-  text: string;
-  textRef: string;
-  id: number;
-};
-type navProps = {
-  isMobile: boolean;
-};
+import { NavDataType, NavPropsType } from "@/components/ui/NavBar/types";
 
-const Navbar = ({ isMobile }: navProps) => {
-  const desktopStyles = "hidden lg:flex w-fit max-h-[20px]";
+import data from "@/data/navbar.json";
+
+const Navbar: FC<NavPropsType> = ({ isMobile }) => {
+  const desktopStyles = "hidden lg:flex w-fit h-fit";
   const mobileStyles = "lg:hidden flex w-fit h-fit mx-auto";
   return (
     <nav className={!isMobile ? desktopStyles : mobileStyles}>
-      <ul className="flex flex-col lg:flex-row gap-6 mx-auto list-none text-center justify-center ">
-        {data.map(({ text, textRef, id }: navData) => (
+      <ul className="flex flex-col lg:flex-row gap-6 mx-auto list-none text-center justify-center">
+        {data.map(({ text, textRef, id }: NavDataType) => (
           <li key={id}>
             <Link
               className="font-sans font-normal text-black transition hover:text-accent focus:text-accent
