+++ src/components/ui/NavBar/types.tsx
@@ -1,0 +1,8 @@
+export interface NavDataType {
+  text: string;
+  textRef: string;
+  id: number;
+}
+export type NavPropsType = {
+  isMobile: boolean;
+};
